+++ src/test/converter2/issues/gh3111.ts
@@ -1,0 +1,7 @@
+export class C {
+    constructor(
+        /** @hidden */
+        readonly x: number,
+        readonly y: boolean,
+    ) {}
+}